fix: useSearchParams와 Suspense 경계 관련 문제에 따른 빌드 실패 해결 (#10)
